requirements.txt
aenum==3.1.15
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
aiosqlite==0.20.0
alembic==1.14.0
annotated-types==0.7.0
anthropic==0.42.0
anyio==4.7.0
asyncpg==0.30.0
attrs==24.3.0
bcrypt==4.0.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
cryptography==44.0.0
dataclasses-json==0.6.7
deepgram-sdk==3.7.7
defusedxml==0.7.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.6
fastapi-cli==0.0.7
frozenlist==1.5.0
fsrs==4.1.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
jwt==1.3.1
langchain==0.3.13
langchain-anthropic==0.3.1
langchain-core==0.3.28
langchain-text-splitters==0.3.4
langsmith==0.2.4
loguru==0.7.3
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.2
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
numpy==2.2.1
orjson==3.10.12
packaging==24.2
passlib==1.7.4
propcache==0.2.1
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.4
pydantic-extra-types==2.10.1
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
pymongo==4.10.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
pytubefix==8.8.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
tenacity==9.0.0
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==13.1
yarl==1.18.3
